+++ model/atmosphere/dycore/tests/dycore_stencil_tests/test_compute_rho_virtual_potential_temperatures_and_pressure_gradient.py
@@ -129,8 +129,8 @@ class TestComputeRhoVirtualPotentialTemperaturesAndPressureGradient(StencilTest)
             dtime=dtime,
             wgt_nnow_rth=wgt_nnow_rth,
             wgt_nnew_rth=wgt_nnew_rth,
-            horizontal_start=int32(0),
+            horizontal_start=0,
             horizontal_end=int32(grid.num_cells),
-            vertical_start=int32(1),
+            vertical_start=1,
             vertical_end=int32(grid.num_levels),
         )
