+++ model/common/src/icon4py/model/common/metrics/stencils/compute_wgtfac_c.py
@@ -50,8 +50,8 @@ def _compute_wgtfac_c(
     k: Field[[KDim], int32],
     nlev: int32,
 ) -> Field[[CellDim, KDim], wpfloat]:
-    wgt_fac_c = where((k > int32(0)) & (k < nlev), _compute_wgtfac_c_inner(z_ifc), z_ifc)
-    wgt_fac_c = where(k == int32(0), _compute_wgtfac_c_0(z_ifc=z_ifc), wgt_fac_c)
+    wgt_fac_c = where((k > 0) & (k < nlev), _compute_wgtfac_c_inner(z_ifc), z_ifc)
+    wgt_fac_c = where(k == 0, _compute_wgtfac_c_0(z_ifc=z_ifc), wgt_fac_c)
     wgt_fac_c = where(k == nlev, _compute_wgtfac_c_nlev(z_ifc=z_ifc), wgt_fac_c)
 
     return wgt_fac_c
